backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
chromadb==0.4.18
pypdf2==3.0.1
python-dotenv==1.0.0
openai==1.3.0
tiktoken==0.5.2